Updated Value Stored in xUnmatched & xMatched
The values stored for the maps "xUnmatched" and "xMatched" were tuple pairs including the index of x and the list of preferences from set Y of x. The index of x was unused; therefore the value of the two maps was simplified to just the list of preferences from set Y of x.